misc/app_testing/parsing_performance/requirements.txt
pika
python-dotenv
python-dateutil
pandas